tools/book-processor/requirements.txt
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
requests>=2.31.0
loguru>=0.7.2
pymupdf
ebooklib>=0.18
beautifulsoup4>=4.12.2
lxml>=4.9.3
